remove used components
